Add files via upload
Welcome to the 'Guess the Number' game! In this exciting challenge, your task is to guess the hidden number chosen by the computer. The game will provide you with clues to help you narrow down your guess. Keep your wits about you, and use your deductive skills to find the correct number in as few attempts as possible. Can you beat the computer and become the ultimate guesser? Let the guessing games begin